Spin image width for clutter box experiment did not allow floating point sizes to be used

+++ src/experiments/clutterBoxExperiment.cpp
@@ -41,7 +41,7 @@ bool contains(std::vector<unsigned int> &haystack, unsigned int needle);
 
 std::vector<unsigned int> computeSearchResultHistogram(size_t vertexCount, const array<ImageSearchResults> &searchResults);
 
-void runClutterBoxExperiment(cudaDeviceProp device_information, std::string objectDirectory, unsigned int sampleSetSize, float boxSize, unsigned int experimentRepetitions, int spinImageWidth) {
+void runClutterBoxExperiment(cudaDeviceProp device_information, std::string objectDirectory, unsigned int sampleSetSize, float boxSize, unsigned int experimentRepetitions, float spinImageWidth) {
 	// --- Overview ---
 	//
 	// 1 Search SHREC directory for files
