+++ src/experiments/clutterBoxExperiment.hpp
@@ -3,4 +3,4 @@
 #include "cuda_runtime.h"
 #include <string>
 
-void runClutterBoxExperiment(cudaDeviceProp device_information, std::string objectDirectory, unsigned int sampleSetSize, float boxSize, unsigned int experimentRepetitions, int spinImageWidth);
+void runClutterBoxExperiment(cudaDeviceProp device_information, std::string objectDirectory, unsigned int sampleSetSize, float boxSize, unsigned int experimentRepetitions, float spinImageWidth);
